player: playerSetMoveBlendFromPlane returns s16
The jump-table addend oracle put a surplus instruction ahead of the
first switch target in two callers; both bound an extsh between the call
and sth r3,1444(r31). Retail stores the return raw into the s16
animEventState field, and MWCC always narrows an int-typed store to a
16-bit lvalue — so the value arrives already narrow and the declared
return type is 16-bit.

playerStateClimbOntoWall 99.429 -> 100.0, playerStateClimbOntoLadder
99.699 -> 100.0, matched_code 2525948 -> 2527372 (+1424), matched
functions 9294 -> 9295. The callee pays 100.0 -> 99.661 for one extsh on
its own return that no in-TU spelling removes: retail has mr r3 in the
callee AND an unnarrowed sth in the caller, which cannot both hold in
one TU — evidence of a real TU boundary here, filed for that lane.

+++ src/dlls/objects/195_Player/player.c
@@ -297,7 +297,7 @@ int playerBuildWallTransitionProbe(GameObject* obj, char* cam, f32* out, f32* ve
 int player_probeClimbable(GameObject* obj, int p4, TrackBBoxHit* src, int dst, int flag);
 int playerStateClimbLedge(int obj, int state, f32 fv);
 int player_SeqFn(int obj, int obj2, ObjSeqState* seq, int endFlag);
-int playerSetMoveBlendFromPlane(int obj, int baseMoveId, int blendMoveId, int* blendAnchor, int* blendPlane,
+s16 playerSetMoveBlendFromPlane(int obj, int baseMoveId, int blendMoveId, int* blendAnchor, int* blendPlane,
                                 f32 samplePhase, f32 moveStepScale, int axis, int flags);
 
 static inline int staffCanContinueSpin(void* state)
@@ -5700,7 +5700,7 @@ void playerStagedRestoreCameraAndSyncPosition(GameObject* obj, int* stateFlags)
     ObjHits_SyncObjectPositionIfDirty(obj);
 }
 
-int playerSetMoveBlendFromPlane(int obj, int baseMoveId, int blendMoveId, int* blendAnchor, int* blendPlane,
+s16 playerSetMoveBlendFromPlane(int obj, int baseMoveId, int blendMoveId, int* blendAnchor, int* blendPlane,
                                 f32 samplePhase, f32 moveStepScale, int axis, int flags);
 
 PlayerModelChainEntry gPlayerModelChainDefault = {lbl_80332EC0, 5};
@@ -10729,7 +10729,7 @@ int playerState00(int obj, int state)
     return 2;
 }
 
-int playerSetMoveBlendFromPlane(int obj, int baseMoveId, int blendMoveId, int* blendAnchor, int* blendPlane,
+s16 playerSetMoveBlendFromPlane(int obj, int baseMoveId, int blendMoveId, int* blendAnchor, int* blendPlane,
                                 f32 samplePhase, f32 moveStepScale, int axis, int flags) {
     ObjModel* model;
     int controlFlags;
